# perf: reduce type resolution to find extension methods for serialization in NetworkBehaviourILPP