This PR will improve the performance of `NetworkBehaviourILPP`.

Currently, `NetworkBehaviourILPP` performs `TypeReference` resolutions for attribute and parameter types to find custom serialization methods in all referenced assemblies. It causes a read to the assembly containing the type if it has not been loaded yet.

We can skip the process earlier by first filtering methods with only the information we can get without `TypeDefinition`.

## Measurement

With `minimalproject`, I used the following code for measurement.

```diff
diff --git a/com.unity.netcode.gameobjects/Editor/CodeGen/NetworkBehaviourILPP.cs b/com.unity.netcode.gameobjects/Editor/CodeGen/NetworkBehaviourILPP.cs
index 7e95fea9..39b4f991 100644
--- a/com.unity.netcode.gameobjects/Editor/CodeGen/NetworkBehaviourILPP.cs
+++ b/com.unity.netcode.gameobjects/Editor/CodeGen/NetworkBehaviourILPP.cs
@@ -77,6 +77,8 @@ namespace Unity.Netcode.Editor.CodeGen
 
             m_Diagnostics.Clear();
 
+            var sw = System.Diagnostics.Stopwatch.StartNew();
+
             // read
             var assemblyDefinition = CodeGenHelpers.AssemblyDefinitionFor(compiledAssembly, out m_AssemblyResolver);
             if (assemblyDefinition == null)
@@ -171,6 +173,8 @@ namespace Unity.Netcode.Editor.CodeGen
 
             assemblyDefinition.Write(pe, writerParameters);
 
+            m_Diagnostics.AddWarning($"ILPP has completed for {compiledAssembly.Name} in {sw.ElapsedMilliseconds} ms");
+
             return new ILPostProcessResult(new InMemoryAssembly(pe.ToArray(), pdb.ToArray()), m_Diagnostics);
         }
 
```

### Before
```
(0,0): warning  - ILPP has completed for Unity.Netcode.Editor in 531 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.EditorTests in 604 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.Components in 363 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.TestHelpers.Runtime in 330 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.RuntimeTests in 735 ms
```

### After
```
(0,0): warning  - ILPP has completed for Unity.Netcode.Editor in 79 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.EditorTests in 128 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.Components in 146 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.TestHelpers.Runtime in 117 ms
(0,0): warning  - ILPP has completed for Unity.Netcode.RuntimeTests in 537 ms
```

## Changelog

- Changed: Improved performance of NetworkBehaviour ILPostProcessor by omitting unnecessary type and assembly resolutions.

## Testing and Documentation

- No tests have been added.
- No documentation changes or additions were necessary.